Check validity of a Metadata Control Number

Checks the format of the Metadata Control Number and warns the user if the provided value might not adhere to the expected format. check_mcn has no access to EPHT systems or APIs and cannot check your mcn against any kind of list of accepted values. It checks if the value is a string, the length of the string and the pattern and nothing else.

Usage

check_mcn(mcn)

Arguments

mcn

Metadata Control Number provided by EPHT.

Value

list containing exit status code and success/failure message

See also

Examples

check_mcn("12345678-1234-1234-1234-123456789012")
#> $code
#> [1] 0
#> 
#> $message
#> Success: mcn
#>